Switched to thread-safe MPI initialisation and fixed race condition in ManyBodyOperator.
MPI has two Init functions: MPI_Init() and MPI_Init_thread(), the latter of which allows us to
request a certain level of thread safety in MPI functions. This is probably not strictly necessary,
since only the master thread calls MPI functions but the MPI runtime is technically allowed to do
weird stuff if we don't use the thread-safe initialiser.