WIP: Replace fine-grained config files with a global struct and key-value interface
This is a pretty big commit that removes my attempt at making a nice,
type-safe specification interface (with various config classes for the
different parts of the code), as C++'s variant types are not expressive
enough to make this work. Instead we just whack the config information
into one global struct, which gets type-checked once during parsing.

I've also added (per Sam/halfflat's prototype) a "map-like" key-value
interface for getting config elements. This lets us retrieve
configuration parameters using the name we give them in the CLI to
do things like spline_type = config_map["Basis/BSpline/SplineType"].
This is *not* type-safe (everything gets converted to/from a
std::any under the hood), but it is a lot more convenient tha directly
referencing the global configuration struct's members.

There's now a proof-of-concept implementation in BasisGenerator which
can serve as a blueprint for porting the rest of the code.