Actually, never mind on the config_map interface. Replacing it with plain references to the gloabl specification struct.
In theory, this is a nice and convenient way to query the user input
parameters. In practice, the underlying map from input keys (strings) to
the actual parameters (could be any type) means we need to store
everything as a `std::any` and defer the type-checking until run-time
and this makes it extremely hard to debug typos and errors. I think
on-balance it's much better for maintainability to have a slightly
less convenient interface that allows the compiler to check types
at build-time.